chore: add the ./dist/ directory to .gitignore
Before using Webpack, if a developer wrote the source code in JavaScript and not in TypeScript,
he had to put the source code directly in ./dist/.

Thus, the ./dist/ directory was kept in the repository.

Now Webpack is being used, no matter if a developer write the source code in TypeScript or in Javascript,
he has to put the source code in ./src-ts/ or in ./src-js/ but never directly in ./dist/,

Thus, the ./dist/ directory should no longer be kept in the repository.